requirements.txt
asgiref>=3.2.5,<3.3.0
Django==3.0.4
django-cors-headers==3.1.0
Pillow==7.0.0
pytz==2019.2
sqlparse==0.3.0
factory-boy==2.12.0
celery==4.4.2
redis==3.4.1
python-dotenv==0.12.0
pipenv==2018.11.26
prospector==1.2.0
psycopg2==2.8.4
djangorestframework==3.11.0
requests==2.22.0
python-dateutil==2.8.1
elasticsearch==7.6.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.6
drf-yasg==1.17.1
stripe==2.45.0